requirements.txt
torch>=2.0,<3.0
transformers>=4.38,<5.0
tokenizers>=0.15
numpy>=1.24
pyyaml>=6.0
tqdm>=4.65
